backend/requirements.txt
fastapi
uvicorn
qrcode[pil]
pillow
python-jose[cryptography]
passlib[bcrypt]
python-multipart